Main function and two others.
I created a main function that stores user data and is connected to MongoDB. It works and sends all the info to our database. In the main.py file you will find two testers I ran as well! I also coded two functions that count the preferences and similarly the overlapping interests.